requirements.txt
dbt-postgres==1.9.*
pandas==3.0.2
xlrd==2.0.2
psycopg2-binary
pydantic-settings
